refactor(engine) Move runtime validation from compiler to engine
