+++ packages/ergo-compiler/lib/apmodelmanager.js
@@ -14,9 +14,11 @@
 
 'use strict';
 
+const slash = require('slash');
 const fsPath = require('path');
 
 const ModelManager = require('@accordproject/concerto-core').ModelManager;
+const ModelFile = require('@accordproject/concerto-core').ModelFile;
 const Builtin = require('./builtin');
 
 /**
@@ -37,6 +39,7 @@ class APModelManager extends ModelManager {
         this.addModelFile(Builtin.ContractModel, '@org.accordproject.contract.cto');
         this.addModelFile(Builtin.RuntimeModel, '@org.accordproject.runtime.cto');
         this.validateModelFiles();
+        this.builtInNamespaces = this.getNamespaces();
     }
 
     /**
@@ -63,6 +66,60 @@ class APModelManager extends ModelManager {
         return models;
     }
 
+    /**
+     * Adds a model file (as a string) to the TemplateLogic.
+     * @param {string} modelFileContent - The model file content as a string
+     * @param {string} fileName - an optional file name to associate with the model file
+     */
+    addAPModelFile(modelFileContent, fileName) {
+        const name = slash(fileName);
+        const modelFile = new ModelFile(this, modelFileContent, name);
+        if (!this.builtInNamespaces.includes(modelFile.getNamespace())) {
+            this.addModelFile(modelFile,name,true);
+        }
+    }
+
+    /**
+     * Add a set of model files to the TemplateLogic
+     * @param {string[]} modelFiles - An array of Composer files as
+     * strings.
+     * @param {string[]} [modelFileNames] - An optional array of file names to
+     * associate with the model files
+     */
+    addAPModelFiles(modelFiles, modelFileNames) {
+        modelFiles.map((modelFileContent, index) => {
+            const modelFileName = slash(modelFileNames[index]);
+            this.addAPModelFile(modelFileContent, modelFileName);
+        });
+        this.validateModelFiles();
+    }
+
+    /**
+     * Update of a given model
+     * @param {string} content - the model content
+     * @param {string} name - the model name
+     */
+    updateModel(content, name) {
+        const currentModels = this.getModelFiles();
+        // Is this a new model?
+        if (!currentModels.some(x => x.getName() === name)) {
+            this.addModelFile(content, name);
+        } else {
+            const previousModelFile =
+                  (currentModels.filter(x => x.getName() === name))[0];
+            const previousContent = previousModelFile.getDefinitions();
+            if (content !== previousContent) {
+                const previousNamespace = previousModelFile.getNamespace();
+                const newNamespace = new ModelFile(this, content, name).getNamespace();
+                if (previousNamespace === newNamespace) {
+                    this.updateModelFile(content, name, true);
+                } else {
+                    this.deleteModelFile(previousNamespace);
+                    this.addModelFile(content, name, true);
+                }
+            }
+        }
+    }
 }
 
 module.exports = APModelManager;
