+++ packages/ergo-engine/package.json
@@ -34,6 +34,7 @@
     "index.browser.js"
   ],
   "dependencies": {
+    "@accordproject/concerto-core": "1.0.0-alpha.6",
     "@accordproject/ergo-compiler": "0.22.0-alpha.4",
     "dayjs": "1.10.4",
     "vm2": "3.5.0"
