+++ packages/ergo-engine/test/data/test3.cto
@@ -1,0 +1,27 @@
+namespace org.accordproject.copyrightlicense
+
+enum Bar {
+  o GRR
+  o YAY
+}
+
+asset Baz identified by bazId {
+  o String bazId
+}
+
+transaction PaymentRequest {
+  o String input
+}
+
+transaction PayOut {
+  o Double amount
+}
+
+concept FOO {
+  o Double amount
+  o Double maybeAmount optional
+  o Integer someNumber
+  o Integer[] someArray
+  o Bar stuff
+  --> Baz relationship
+}
